+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Added
 
+- **! Added the `--only=<name>` option, which will only run that specific formatter/linter.**
+- **Added names to formatters and linters.**
 - **Added support for python 3.6**
 
 ## Changed